Move the engine to single-precision math.
id's C calls libm's double sin/cos/sqrt/floor/etc. and writes its float
constants unsuffixed, and the EE has no double FPU: every one of those ran
through libgcc's soft-float routines. In the perf demo that was the whole
cause of the dropped frames - particle effects doing double trig per particle
per frame (flies), and temp entities doing it per world unit of a rail trail
at parse time.

- -fsingle-precision-constant for the engine C. The C++ backend already
  suffixes its constants and -Wdouble-promotion keeps it that way.
- ps2/math/math_c.h: the scalar functions the engine and ps2::math share,
  static inline with C linkage, C89 that also passes the backend's C++
  warnings. 129 engine calls in 26 files now use them; tan/atan/atan2 (15
  calls, none hot) take newlib's float versions. math.h wraps them and
  math.cpp is gone.
- Cosf/Sinf/ASinf rewritten from the Morten Mikkelsen asm into plain C, so
  inlined calls load their constants once per loop, interleave, and drop when
  unused: Cosf 50 -> 35 instructions standalone and ~21 per result in a loop,
  at slightly better accuracy (6e-7 within a turn, minimax degree 9). ASinf
  58 -> 42 instructions and 8e-4 -> 3e-6 max error - the asm was a truncated
  Taylor series - and no longer writes $8-$11 without declaring them
  clobbered. Fmodf no longer returns -y for an exact multiple of y.

Perf demo (7677 frames) against the capture before this: frames missing
vsync 59 -> 1, EE work mean 7.9 -> 5.4 ms and p99 16.2 -> 9.5 ms, ClScene
-83%, ClParse -74%, SndMix -9%, rail trail parse 34 -> 2.5 ms. The one drop
left is a mid-demo asset load, not math.

+++ src/client/cl_view.c
@@ -428,8 +428,8 @@ float CalcFov(float fov_x, float width, float height)
         Com_Error(ERR_DROP, "Bad fov: %f", fov_x);
     }
 
-    x = width / tan(fov_x / 360 * M_PI);
-    a = atan(height / x);
+    x = width / tanf(fov_x / 360 * M_PI);
+    a = atanf(height / x);
     a = a * 360 / M_PI;
 
     return a;
